adds logging api module
This commit is the initial add of the logging api module used to configure
the global system logging options.  This basic modules provides the ability
to enable or disable global logging and insert or remove destination hosts.
This update includes the system spec tests for use with the logging
module