adds communities and notifications support to snmp module
This commit adds support to the snmp module for parsing and configuring snmp
communities and notifications.   Communities provides a new hash in the
resource that include the community name, access type and acl.  The
communities hash is indexed by community name.  Notifications provides a new
array in the resource hash that includes the the list of notifications.  Each
notificaiton provides the name and state of the notification from the nodes
running configuration