adds new methods to the Entity class
This commit adds two new instance methods to the Entity class to simplify
sending configuration commands

The first method is configure_interface which takes an interface name and
a set of commands to configure the interface with.

The second method is command_builder which will take a command string and
opts specify value and default.  The command_builder will return the
command with appropriate keywords and/or value substitutions