Eliminate overloading value option in command_builder.
Command builder currently overloads the value parameter. When the value
is set it is used as a value in building the command. When the value is
false then the command is negated. This doesn't allow a value to be
specified when the command is negated.

Not all the api modules are using command builder so they were updated to do so.

The advantage of having all the simple command methods that use
default/no is that the same library is used consistently throughout the
modules reducing duplicated code patterns.

Test cases for no/default cases of set methods were added.